Move action to own file
