+++ jest.config.js
@@ -1,7 +1,5 @@
 module.exports = {
   clearMocks: true,
-  moduleFileExtensions: ["js", "ts"],
   testEnvironment: "node",
-  testMatch: ["**/*.spec.js"],
   verbose: true
 };
